requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
Click==7.0
cloudpickle==0.5.6
cryptography==2.3.1
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
Flask==1.0.2
future==0.16.0
idna==2.7
ItsDangerous==1.0.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.0
networkx==2.2
numpy==1.15.1
pandas==0.23.4
Pillow==5.2.0
pycparser==2.19
pyparsing==2.2.1
python-dateutil==2.7.3
python-telegram-bot==11.1.0
pytz==2018.5
PyWavelets==1.0.0
redis==2.10.6
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
Werkzeug==0.14.1
